requirements.txt
aws-cdk.core==1.69.0
aws-cdk.aws-lambda==1.69.0
aws-cdk.aws-cloudformation==1.69.0
aws-cdk.aws-s3==1.69.0
aws-cdk.aws-iam==1.69.0
aws-cdk.aws-codedeploy==1.69.0
aws-cdk.aws-codebuild==1.69.0
aws-cdk.aws-codecommit==1.69.0
aws-cdk.aws-codepipeline_actions==1.69.0
aws-cdk.aws-secretsmanager==1.69.0
aws-cdk.aws-ssm==1.69.0